bank/requirements.txt
Flask==2.1.3
PyJWT==2.4.0
python-jose==3.3.0  # For JWS signing
